requirements.txt
fastapi
uvicorn[standard]
aio-pika
aiosmtplib
python-dotenv
